:sparkles: Handle more keys
